@@ -1,7 +1,7 @@
 require 'data_helpers'
 
 class Admin::BaseController < ApplicationController
-  before_filter :signed_in_user, :require_admin
+  before_action :signed_in_user, :require_admin
 
   def require_admin
     unless current_user.can_admin?
